Update enclosure files
Add door frame arms to hinges - its hard to fit on a single plater but
it is just possible and this hopefully makes the dppr frame les inclined
to fall apart.

Also change the thickness of the perspect doors it was too large which
was probably the main reason the parts kept falling off.

Also use larger magnets so the doors stay closed.