We need a second i2c bus!
Multi-master on i2c should in theory work but it looks like the hardware
and software is pretty much configured to assume one master out of the
box.

To avoid the complications the easiet way forward is to add a 2nd i2c
port which is a slave.

So we have:

|RPI octoprint| -> i2c 0 -> |driver board| -> i2c 1 -> |tmp117|

The slave i2c port can be used to set the set point, configure the PID
parameters and read the current value of the TMP117.